Merge py2 and py3 branch to a single branch.
This integrates the py2 and py3 branches to a single branch with explicit version checking where necessary (in particular in the places where we need to know if we're working with bytes or actual strings).

Most future fixes should hopefully don't have to touch this. If we find that future maintenance require extending our current testing, we might want to wrap certain string handling in `six` or a similar library.

I do however think that we'll go with just py3.5 at some time in the future, finally deprecating the 2.7 version.