Handle URLs as string, pick random URL for fetch
This patch allows a single URL to be provided to the __init__ method for
the client.

    Client('http://imbo.foo', ...)  # this now works

The previous handling would try to extract a URL from each character in
the string, making it hard to debug why your client tried to access the
host `i`.

The second change uses random.choice() to pick one of the supplied urls,
spreading the requests randomly across the defined urls. The next patch
will introduce a setting and update the method that picks hosts from a
image_id.