Clean up exception messages
We need to explicitly cast these exceptions to a str before using them - as
we're concatenating the messages.

The ValueError is thrown by the JSON parsing, so the response line should
be moved outside of the try/catch statement, so that we can be sure that
the `response` value is actually defined for the catch statement.