Avoid assigning an object property when testing
There is no need to assign these client tests to the object itself, as the
result is tested within the method. The assignment _could_ have side
effects somewhere later (if the teardown for some reason isn't ran as
assumed).

To avoid any potential side effects we keep the assignments local to the
method.